Added trap movement

+++ src/Circle.java
@@ -6,7 +6,7 @@ import java.awt.*;
 public class Circle extends Entity {
 
 
-    public Circle(Game game, int x, int y, int width, int height, double dx, double dy, Color color) {
+    public Circle(Game game, double x, double y, double width, double height, double dx, double dy, Color color) {
         super(game, x, y, width, height, dx, dy, color);
     }
 
@@ -14,7 +14,7 @@ public class Circle extends Entity {
     public void paint(Graphics g) {
 
         g.setColor(getColor());
-        g.fillOval(getX(),getY(),getWidth(),getHeight());
+        g.fillOval((int)getX(),(int)getY(),(int)getWidth(),(int)getHeight());
 
     }
 }
