+++ src/Trap.java
@@ -6,7 +6,7 @@ import java.awt.*;
 public class Trap extends Entity {
 
 
-    public Trap(Game game, int x, int y, int width, int height, double dx, double dy, Color color) {
+    public Trap(Game game, double x, double y, double width, double height, double dx, double dy, Color color) {
         super(game, x, y, width, height, dx, dy, color);
     }
 
@@ -14,7 +14,7 @@ public class Trap extends Entity {
     public void paint(Graphics g) {
 
         g.setColor(getColor());
-        g.fillRect(getX(),getY(),getWidth(),getHeight());
+        g.fillRect((int)getX(),(int)getY(),(int)getWidth(),(int)getHeight());
 //        int[] xPoints = {getX(),getX()+20,getX()+10};
 //        int[] yPoints = {getY(),getY(),getY()+20};
 //        g.fillPolygon(xPoints,yPoints,3);
